Reorganized tests folder, modularized front end tests
