+++ jest.config.js
@@ -185,8 +185,10 @@ module.exports = {
   // watchman: true,
   transform: {
     '^.+\\.js$': '<rootDir>/node_modules/babel-jest',
+    //'.*\\.(vue)$': '<rootDir>/node_modules/vue-jest',
     '.*\\.(vue)$': '<rootDir>/node_modules/jest-vue-preprocessor',
-    '^.+\\.vue$': 'vue-jest',
+    '.*\\.(vue)$': '<rootDir>/node_modules/vue-jest',
+    //'^.+\\.vue$': 'vue-jest',
     '.+\\.(css|styl|less|sass|scss|svg|png|jpg|ttf|woff|woff2)$': 'jest-transform-stub',
     '^.+\\.jsx?$': 'babel-jest'
   },
