+++ tests/updateMockData.js
@@ -1,0 +1,11 @@
+/**
+* @Author: Milan Donhowe
+* @Email: Milan.Donhowe@oregosntate.edu
+* @Date Created:       4/20/2021
+* @Date Last Modified: 4/20/2021
+* @Description: Dynamically update the Mock Data
+*/
+
+const fs = require('fs')
+
+// TODO query api & build new mock data
